[ORB-00330] Split Rust extractor module
